Code Improvements
Added code to double-check and make sure that the correct time server script installed by ntpMerlin (for ntpd or chronyd) is found and launched. This check is performed at startup following a reboot, every time the time server is restarted, and every 10 minutes via the pre-existing cron job to generates time server stats.